requirements.txt
Django==3.2
# Django==4.1.5
gunicorn==20.1.0
python-dotenv==0.20.0
# mssql-django==1.1.3
mssql-django==1.2
pyodbc==4.0.34
whitenoise==6.2.0
djangorestframework==3.13.1
pyjwt==2.4.0
django-crispy-forms==1.14.0
django-cors-headers==3.13.0
Pillow==9.4.0






